# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Checkout API Documentation",
        "description": "Checkout API Swagger OpenApi",
        "contact": {
            "email": "jpedro747.jp@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/coupons/{value}": {
            "get": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Obter cupons pelo nome",
                "description": "Encontra cupom pelo nome",
                "operationId": "d58ddd79d979605c72fafb1f3a3bee08",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "O nome do cupom para verificar dados do cupom",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "OFF10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do cupom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coupon"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/coupons-validate": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Validar um cupom",
                "description": "Verifica se um cupom é válido com base em seu código, plano associado, expiração e limites de uso.",
                "operationId": "f12b3be9ca3540427c60ea2d74c1bd7a",
                "requestBody": {
                    "description": "Cupom e ID do Plano para validar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CouponValidationPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cupom é válido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CouponValidationSuccess"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação (Ex: cupom não existe) ou cupom inválido (Ex: expirado).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ValidationErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/CouponValidationFailure"
                                        }
                                    ]
                                },
                                "examples": {
                                    "validationError": {
                                        "summary": "Erro de Validação (Laravel $validate)",
                                        "value": {
                                            "message": "The data provided is invalid.",
                                            "errors": {
                                                "coupon": [
                                                    "Invalid coupon!"
                                                ]
                                            }
                                        }
                                    },
                                    "couponInvalid": {
                                        "summary": "Erro de Regra (Serviço)",
                                        "value": {
                                            "valid": false,
                                            "message": "Expired coupon."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Listar todas as assinaturas",
                "description": "Retorna uma lista de todas as assinaturas com seus relacionamentos.",
                "operationId": "a57381ef1031fcbb133278a38ec4138c",
                "responses": {
                    "200": {
                        "description": "Lista de assinaturas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubscriptionWithRelations"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Criar uma nova assinatura (Checkout - Passo 1)",
                "description": "Recebe o plano, cliente e cupom. Valida o cupom e cria a assinatura inicial (ainda inativa).",
                "operationId": "0fed2b35010e282f6ccc656ac915f78b",
                "parameters": [
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "Uma chave única (UUID) usada para garantir que a requisição seja processada apenas uma vez.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a1b2c3d4-5678-90ab-cdef-1234567890ab"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Detalhes da assinatura",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubscriptionPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assinatura criada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação (Ex: email duplicado) ou erro de regra (Ex: cupom inválido).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ValidationErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SimpleErrorResponse"
                                        }
                                    ]
                                },
                                "examples": {
                                    "validationError": {
                                        "summary": "Erro de Validação (Laravel $validate)",
                                        "value": {
                                            "message": "The data provided is invalid.",
                                            "errors": {
                                                "email": [
                                                    "This email already contains a signature. Please use a different email address.."
                                                ]
                                            }
                                        }
                                    },
                                    "couponError": {
                                        "summary": "Erro de Regra (Serviço/Exception)",
                                        "value": {
                                            "message": "Expired coupon."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/subscriptions/{id}": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Buscar uma assinatura específica",
                "description": "Retorna os dados de uma assinatura e seus relacionamentos.",
                "operationId": "9f5f8baf2f2e4388c9751660f0e683dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Assinatura",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da assinatura",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionWithRelations"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Assinatura não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimpleErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CouponValidationPayload": {
                "title": "Coupon Validation Payload",
                "description": "Payload necessário para validar um cupom",
                "required": [
                    "coupon"
                ],
                "properties": {
                    "coupon": {
                        "description": "O código do cupom para validar",
                        "type": "string",
                        "example": "OFF10"
                    },
                    "plan_id": {
                        "description": "O ID do plano ao qual este cupom será aplicado",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CouponValidationSuccess": {
                "title": "Sucesso na Validação do Cupom",
                "description": "Resposta de sucesso para um cupom válido",
                "properties": {
                    "valid": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Valid Coupon!"
                    }
                },
                "type": "object"
            },
            "CouponValidationFailure": {
                "title": "Falha na Validação do Cupom",
                "description": "Resposta de falha para um cupom inválido (regra de negócio)",
                "properties": {
                    "valid": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Expired coupon."
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "title": "Resposta de Erro de Validação",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The data provided is invalid."
                    },
                    "errors": {
                        "description": "Objeto onde as chaves são os nomes dos campos e os valores são arrays de mensagens de erro.",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "SubscriptionWithRelations": {
                "title": "Assinatura com Relações",
                "description": "Assinatura com todos os detalhes relacionados (plano, cupom, transação)",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Subscription"
                    },
                    {
                        "properties": {
                            "plan": {
                                "$ref": "#/components/schemas/Plan"
                            },
                            "coupon": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Coupon"
                                    }
                                ],
                                "nullable": true
                            },
                            "transaction": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "SimpleErrorResponse": {
                "title": "Resposta de Erro Simples",
                "description": "Resposta de erro simples",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message."
                    }
                },
                "type": "object"
            },
            "StoreSubscriptionPayload": {
                "title": "Payload de Criação de Assinatura",
                "description": "Payload completo para o endpoint de checkout.",
                "required": [
                    "plan_id",
                    "email"
                ],
                "properties": {
                    "coupon": {
                        "type": "string",
                        "example": "OFF10",
                        "nullable": true
                    },
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    }
                },
                "type": "object"
            },
...
            "Coupon": {
                "title": "Coupon",
                "description": "Discount Coupon Model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The coupon code",
                        "type": "string",
                        "example": "SAVE30"
                    },
                    "plan_id": {
                        "description": "Specific plan ID (null for global)",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "expiration_days": {
                        "description": "Days to expire (null for never)",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "amount_of_uses": {
                        "description": "Total number of uses allowed",
                        "type": "integer",
                        "example": 2
                    },
                    "discount_percent": {
                        "type": "number",
                        "format": "float",
                        "example": 20,
                        "nullable": true
                    },
                    "discount_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 30,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Plan": {
                "title": "Plan",
                "description": "Plan Model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the plan",
                        "type": "string",
                        "example": "PRO_MONTHLY"
                    },
                    "description": {
                        "description": "Description of the plan",
                        "type": "string",
                        "example": "Pro Anual"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.9
                    },
                    "periodicity": {
                        "description": "Plan duration in days",
                        "type": "integer",
                        "example": 30
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Subscription": {
                "title": "Subscription",
                "description": "Subscription Data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "coupon_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "title": "Transaction",
                "description": "Transaction Model (Payment Record)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "card_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "coupon_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "subscription_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "price_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 99.9
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "card": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Card"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}